Created CartContext, useCart and useLocalStorage
Created CartContext.jsx, useCart.jsx and useLocalStorage.jsx, and implemented them so that the product is being stored when Add to Cart on ProductPage is pressed.

I also created the CartBadge to show the user how many items there is in cart.

Product component, CartPage and clearing of cart is next up