Created "Added"
Created an "Added" state to the Add to cart button on Product. Now, it's not possible to spam the button and add more thatn one priduct at every refresh.

The user can add more of the same item by either visiting the same product multiple times, or using the "+" in the CheckoutPage